Fixed #define for non WIndows
git-svn-id: svn://svn.code.sf.net/p/mtsieve/svn@25 bb85bfea-1d01-4b18-b6d8-510d50b5ba24

+++ dm_divisor/DMDivisorApp.cpp
@@ -32,7 +32,7 @@
 #ifdef WIN32
 static const unsigned long long _MONTGOMERY_DATA[14] = {1,0,0,0,0,0,0,0,0,0,0,0,0,UINT64_C(1)<<(1<<PRE_SQUARE)};
 #else
-static const unsigned long long _MONTGOMERY_DATA[14] = {1,0,0,0,0,0,0,0,0,0,0,0,0,UINT64_C(1)<<(1<<PRE_SQUARE)};
+static const unsigned long _MONTGOMERY_DATA[14] = {1,0,0,0,0,0,0,0,0,0,0,0,0,UINT64_C(1)<<(1<<PRE_SQUARE)};
 #endif
 
 #define ONE    (_MONTGOMERY_DATA+0)
